requirements.txt
# Core dependencies for Trello AI with LangChain
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# LangChain for agent functionality
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-core>=0.1.0

# OpenAI for GPT-4 integration
openai>=1.0.0

# Additional utilities
typing-extensions>=4.0.0
uuid